Add a declaration for FCM to improve type autocompletion
