--- core/interfaces/firebase.ts
@@ -1,31 +1,0 @@
-interface FirebaseKeys {
-    privateKey: string
-    publicKey: string
-    authSecret: string
-}
-
-interface FirebasePushCredential {
-    token: string
-    pushSet: string
-}
-
-interface LegacyPushCredential {
-    token: string
-    androidId: string
-    securityToken: string
-    appId: string
-}
-
-export interface FirebaseCredentials {
-    keys: FirebaseKeys
-    fcm: FirebasePushCredential
-    gcm: LegacyPushCredential
-}
-
-export interface PushData {
-    readonly from: string
-    readonly priority: string
-    readonly title: string
-    readonly message: string
-    readonly reserved: string
-}
